requirements.txt
torch
torchvision
numpy
tqdm
sklearn
scikit-image
scikit-learn
simpleitk
scipy
timm
Pillow
pretrainedmodels
efficientnet_pytorch